Improve code organization and documentation for ruby environment.
